First steps to actual refactoring (#7)
Intorduces new class MkxFile(), which started as a copy of Utils(). No code actually uses it, for
now. Just wanted to get this in as a preview.
Also, some minor change: do not rename os.path, found it misleading. I'd expect an actual path to
be stored under such a name.