More efficiency, hopefully (#7)
I consider this a semi-major change. Prior to this we always created two instances of
MatoroskaFile(), one while filling utils.files and checking if it is a matroska and then when
actually processing in process_thread(). But at that stage we have all the information we need
already. So pass the path of the file, which is guaranteed to be a matroska into the queue. Since
utils.files has now become an ordered dictionary, it also holds an entire MatroskaFile object with
all the methods attached. Before we just threw away that object when we only tried initializing the
object and then only take its path and put that in utils.files.
While that might hurt memory usage I think it is worth it for not having to do the checking twice.