mkvrg: More robust track handling
Matroska files can have tracks in arbitrary order, which made mkvrg fail
with some files that did not have tracks in the expected order. This
adds a dependency for ffprobe, which should already be installed
alongside ffmpeg in most cases.

Also, minor cosmetics to the description which now reflects the actual
requirements that are checked.

Fixes #11