Don't forcibly remove files.
When in non-interactive mode, i.e. in a script, aliases are not
expanded. Aliases like `alias rm='rm -i'` are the primary reason for the
inflationary, and ill-advised, use of `rm -f`, which has more
implications than "overriding" `-i`. It basically does what it
advertises and, in the process, shuts up any warning that might be
useful in some corner cases.