More slashing.
Removed all filters but 'ebur128'. There is no point in using the other
two. 'replaygain' is apparently the (deprecated) 1.0 version, judging by
a quick glance at the docs and code. And since RG 2.0 is based on
loudness measurement according to BS.1770/EBU-R128, only with a
different reference loudness, the 'ebur128' provides all we need.
'loudnorm' uses the exact same algorithm but unnecessarily slows things
down by always running a dynamic output normalization (which gets send
to the null device in our case) with 192kHz sample rate, no less--talk
about overkill (someone should tell the author about the limits of human
hearing). No wonder it is that slow. In my tests the results were the
same between ebur128 and loudnorm, give or take the odd rounding
difference of no more than 0.1 dB, which is imperceptible. So, long
story short, 'ebur128' it is.

Also, the reference level is now -18 dBFS/LUFS by default, and the help
for the new '-R' option, which "replaces" OFFSET, contains a strong
discouragement of changing it. In short, if the tag says REPLAYGAIN_*,
then the reference should be that of RG 2.0. This change also simplifies
parsing and provides for the earliest possible exit in case of invalid
values. BTW, fractional values, let alone *two* decimal digits, make
little sense in this context, which is why the docs contain integers
now, although one can still use floats.

Further, all but the FFMPEG environment variables will be ignored now,
since they can be set by command line options.