Logging cosmetics.
Don't output too much info in the summary. Especially the 'gain' was
confusing; essentially, when applied the value should be the reference
loudness, but previous output was reversal of calculation from the
parsed measurements, and, since the reference is well documented, there
is no point in stating the obvious.

If more information is desired, there is always the debug log level.